Migrated UCI to NextJs
