+++ .env.local
@@ -1,0 +1,15 @@
+NEXT_PUBLIC_MAGICBELL_API_KEY="$MAGICBELL_API_KEY"
+NEXT_PUBLIC_MAGICBELL_USER_EMAIL="$MAGICBELL_USER_EMAIL"
+NEXT_PUBLIC_TRANSPORT_SOCKET_URL="wss://comms-playground.samagra.io"
+
+
+# FireBase credentials
+NEXT_PUBLIC_FIREBASE_API_KEY="AIzaSyBAW9eosfF-oo5xPXXeoBp810YLyFlYZJw"
+NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN="testing-uci-web-channel.firebaseapp.com"
+NEXT_PUBLIC_FIREBASE_PROJECT_ID="testing-uci-web-channel"
+NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET="testing-uci-web-channel.appspot.com"
+NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID="584834206650"
+NEXT_PUBLIC_FIREBASE_ID="1:584834206650:web:b1a8a5a03dd231406e2b83"
+NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID="G-WSYRRR972G"
+
+NEXT_PUBLIC_FIREBASE_VAPID_KEY="BKuweez_2-hT2OYFRwUZ6CUuH3B3wrY0y1EOWlnbYtML0Iy7okQQZiiFZwzIYoin5TqdglT-18k-ki5cY_aUyjc"
