--- .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,30 +1,0 @@
----
-name: Bug report
-about: Create a report to help us improve
-labels: bug
----
-
-<!-- Before filling this issue, please read the wiki (https://github.com/Samagra-Development/uci-pwa/wiki)
-and search if the bug do not already exists in the issues (https://github.com/Samagra-Development/uci-pwa/issues). -->
-
-### Description
-
-Please describe your issue in few words here.
-
-#### How to reproduce
-
-Describe the bug and list the steps you used when the issue occurred.
-
-#### Screenshots
-
-If applicable, add screenshots to help explain your problem.
-
-### Versions
-
-* Last commit id on master:
-* Operating System (lsb_release -a):
-* Browser [e.g. chrome, safari]:
-
-### Logs
- 
-Any logs (if any) generated in
