--- .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,20 +1,0 @@
----
-name: Feature request
-about: Request a new feature in UCI-PWA
-labels: enhancement
----
-
-<!-- Before filling this issue, please read the Wiki (https://github.com/Samagra-Development/uci-pwa/wiki)
-and search if the bug do not already exists in the issues (https://github.com/Samagra-Development/uci-pwa/issues). -->
-
-### Description
-
-Please describe your situation in few words here.
-
-#### Steps followed and expected result
-
-Describe the steps followed by you and your expected results after following the steps.
-
-#### Screenshots
-
-If applicable, add screenshots to help explain your problem.
