--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,18 +1,0 @@
-<!-- Please refer to CONTRIBUTING.md (https://github.com/Samagra-Development/uci-pwa/blob/main/CONTRIBUTING.md)
-before creating the pull request to make sure you follow all the standards. -->
-
-## Description
-
-Please describe the changes in your pull request in few words here.
-
-### Changes
-
-List the changes done to fix a bug or introducing a new feature.
-
-## How to test
-
-Describe the steps required to test the changes proposed in the pull request.
-
-Please consider using the closing keyword if the pull request is proposed to
-fix an issue already created in the repository
-(https://help.github.com/articles/closing-issues-using-keywords/)
