--- .prettierignore
@@ -1,17 +1,0 @@
-# Dependency directories
-node_modules/
-
-# dependencies
-package.json
-package-lock.json
-
-# testing
-/coverage
-
-# production
-/build
-
-.github/
-/public
-.min.js
-.min.css
