--- .prettierrc.json
@@ -1,6 +1,0 @@
-{
-  "semi": true,
-  "singleQuote": false,
-  "tabWidth": 2,
-  "useTabs": false
-}
