+++ package.json
@@ -1,13 +1,20 @@
 {
-  "name": "ui-pwa",
+  "name": "next-uci-web-channel",
   "version": "0.1.0",
   "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
+  },
   "dependencies": {
     "@chakra-ui/icons": "^1.1.5",
     "@chakra-ui/react": "^1.8.3",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
     "@magicbell/magicbell-react": "^8.5.3",
+    "@storybook/react": "^6.5.9",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^10.4.9",
     "@testing-library/user-event": "^12.8.3",
@@ -16,53 +23,20 @@
     "@types/react": "^16.14.23",
     "@types/react-dom": "^16.9.14",
     "bootstrap": "^5.1.3",
-    "framer-motion": "^4.1.17",
-    "react": "^17.0.2",
+    "framer-motion": "^6.3.15",
+    "next": "12.1.6",
+    "react": "18.2.0",
     "react-alert": "^7.0.3",
     "react-alert-template-basic": "^1.0.2",
     "react-bootstrap": "^2.4.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "18.2.0",
     "react-icons": "^4.3.1",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
-    "react-toastify": "^9.0.1",
-    "socket.io-client": "^4.4.1",
+    "react-toastify": "^9.0.5",
+    "socket.io-client": "^4.5.1",
     "typescript": "^4.5.5",
-    "web-vitals": "^0.2.4"
-  },
-  "scripts": {
-    "start": "set BROWSER=none && react-scripts start",
-    "build": "set CI='' && react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "format": "prettier --write .",
-    "format:check": "prettier --check .",
-    "storybook": "start-storybook -p 6006 -s public",
-    "build-storybook": "build-storybook -s public"
-  },
-  "eslintConfig": {
-    "extends": "react-app",
-    "overrides": [
-      {
-        "files": [
-          "**/*.stories.*"
-        ],
-        "rules": {
-          "import/no-anonymous-default-export": "off"
-        }
-      }
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
     "@cypress/react": "^5.12.5",
@@ -82,6 +56,11 @@
     "cypress": "^9.7.0",
     "prettier": "^2.5.1",
     "webpack": "^5.69.1",
-    "webpack-dev-server": "^4.9.0"
+    "webpack-dev-server": "^4.9.0",
+    "@types/node": "18.0.0",
+    "@types/react-dom": "18.0.5",
+    "eslint": "8.18.0",
+    "eslint-config-next": "12.1.6",
+    "typescript": "4.7.4"
   }
 }
