--- renovate.json
@@ -1,25 +1,0 @@
-{
-  "extends": ["config:base"],
-  "automerge": true,
-  "prHourlyLimit": 2,
-  "updateNotScheduled": false,
-  "timezone": "Asia/Calcutta",
-  "schedule": ["after 10pm and before 5am on every weekday", "every weekend"],
-  "masterIssue": true,
-  "labels": ["type: dependencies", "renovate"],
-  "packageRules": [
-    {
-      "packagePatterns": ["*"],
-      "excludePackagePatterns": [
-        "react",
-        "react-dom",
-        "react-scripts",
-        "@cypress/react",
-        "@cypress/webpack-dev-server",
-        "cypress",
-        "webpack-dev-server"
-      ],
-      "enabled": false
-    }
-  ]
-}
