--- src/App.spec.tsx
@@ -1,8 +1,0 @@
-import * as React from "react";
-import { mount } from "@cypress/react";
-import App from "App";
-
-it("renders the home page", () => {
-  mount(<App />);
-  cy.contains(/chaks/i);
-});
