+++ src/components/TextBar/index.tsx
@@ -3,7 +3,6 @@ import { Box, Button, Input } from "@chakra-ui/react";
 import { MdSend } from "react-icons/md";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import "styles/global.css";
 
 
 const TextBar = (props: any) => {
