+++ src/components/websocket.ts
@@ -1,10 +1,11 @@
-const host =
-  process.env.NODE_ENV === "production"
-    ? window.location.host
-    : "localhost:3005";
-const { io } = require("socket.io-client");
+import {io} from 'socket.io-client';
+// const host =
+//   process.env.NODE_ENV === "production"
+//     ? window.location.host
+//     : "localhost:3005";
 
-export const socket = io(`${process.env.REACT_APP_TRANSPORT_SOCKET_URL}`);
+
+export const socket = io(`${process.env.NEXT_PUBLIC_TRANSPORT_SOCKET_URL}`);
 export const send = (msg: any, session: any) =>
   socket.emit("botRequest", {
     content: {
